requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sarvamai>=0.1.0
chromadb==0.5.0
sentence-transformers==3.0.0
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.9.0
apscheduler==3.10.4
numpy>=1.26.0
